# proc: only format registers value when it's necessary