… pkg/proc/core/core_test.go
… pkg/proc/winutil/regs.go
… service/api/conversions.go