@aarzilli 2020-02-12
I realize that the amd64DwarfToName map isn't used anymore, except to build it's inverse map (amd64NameToDwarf) and so we could remove. However I'd like to keep it around since it will eventually be used to implement #1838 (we'll need a way to put a name on registers for which we only have a dwarf register number).

Also, I explored the possibility of a much more radical change where the backend specifies the DWARF register number directly (instead of the thing we have now where we find the register number by using the register name). Technically it's possible, however for the gdbserial backend we only have the register names so it would simply push the ugly (register name → register number) table down, so it isn't really worth it.

Finally the extra register names that get added manually in the constructor of amd64NameToDwarf are alternate names for some registers that mozilla rr uses.